split out context
